Move GeneralTab to gui.preftabs
