+++ src/main/java/net/sf/jabref/PrefsDialog3.java
@@ -45,6 +45,7 @@ import net.sf.jabref.gui.FileDialogs;
 import net.sf.jabref.gui.MainTable;
 
 import com.jgoodies.forms.builder.ButtonBarBuilder;
+import net.sf.jabref.gui.preftabs.GeneralTab;
 import net.sf.jabref.gui.preftabs.PrefsTab;
 import net.sf.jabref.util.Util;
